Create README.md
You are welcome to design your own board. We do not allow 3rd party boards to use our name or logo, but you are welcome to indicate "BrainBoardz Compatible" on your board if it is 100% compatible and it is manufactured using a Lead Free process.  Please ensure that you follow our design guidelines very closely. The pin assignments have to be 100% compatible and the connector placement is extremely critical.  Also, please ensure that Neurons run properly (including the SD-card, Blinky and the Piezo) on your design.  These schematics above are being released under the Creative Commons Attribution Share-Alike license Version 4.0.